+++ src/repositories/calculator_repository.py
@@ -17,11 +17,12 @@ class CalculatorRepository:
         db = self._connection.cursor()
         db.execute("SELECT * FROM Calculations")
         rows = db.fetchall()
+        print(rows)
         return rows
 
     def delete_calculations(self):
         db = self._connection.cursor()
-        db.execute("DROP TABLE Connections")
+        db.execute("DROP TABLE IF EXISTS Calculations")
 
 
 calculator_repository = CalculatorRepository(get_database_connection())