+++ src/services/about_service.py
@@ -1,0 +1,10 @@
+from ui.about_view import about_view as default_about_view
+
+class AboutService:
+    def __init__(self, about_view = default_about_view):
+        self._about_view = default_about_view
+        
+    def initialize_about_view(self):
+        self._about_view.show_about_view()
+    
+about_service = AboutService()